deprecated: deprecate more schemas
